requirements_osx.txt
anaconda-client==1.9.0
anaconda-project==0.10.2
conda==4.14.0
torch==1.11.0
scikit-learn==1.0.2
matplotlib==3.5.1
numpy==1.21.5
scipy==1.7.3
pandas==1.3.5
tqdm>=4.62.3,<5.0.0
nltk==3.7
gensim==4.1.2
joblib==1.1.0
lxml==4.8.0
seaborn==0.11.2
scispacy==0.5.1
spacy==3.2.4
transformers==4.21.2
python-dotenv==0.21.0
requests==2.27.1
sacred==0.8.2
neptune-client==0.16.8
neptune-sacred==0.10.0